CMake: Do not use CMAKE_LIBRARY_ARCHITECTURE for linux
CMAKE_LIBRARY_ARCHITECTURE is engineered for Debian-derived linux,
not for other distros such as RHEL-derived, Arch, etc.

To detect compilers in GetCompilerAndPlatform, use instead
`CMAKE_CXX_COMPILER_ID` and `CMAKE_CXX_COMPILER`, as it is used
already for the `windows` case.

Note that the arm version is untested, so a double check with
COMPILER_ID and just COMPILER is in place.

Fixes #39

Signed-off-by: ns <portification@gmail.com>